requirements.txt
discord.py==2.3.2
uvicorn==0.29.0
fastapi==0.115.4
flask
pytz
python-dotenv